(IPU) Backport - Store thresholds for color conversions in u16

+++ pcsx2/IPU/IPU.cpp
@@ -1,5 +1,5 @@
 /*  PCSX2 - PS2 Emulator for PCs
- *  Copyright (C) 2002-2010  PCSX2 Dev Team
+ *  Copyright (C) 2002-2022  PCSX2 Dev Team
  *
  *  PCSX2 is free software: you can redistribute it and/or modify it under the terms
  *  of the GNU Lesser General Public License as published by the Free Software Found-
@@ -36,7 +36,7 @@ __aligned16 decoder_t decoder;
 
 // Quantization matrix
 static rgb16_t vqclut[16];			//clut conversion table
-static u8 s_thresh[2];				//thresholds for color conversions
+static u16 s_thresh[2];				//thresholds for color conversions
 int coded_block_pattern = 0;
 
 alignas(16) static u8 indx4[16*16/2];
